Fix grossly misleading information
This patch has been provided by an anonymous user and solves an issue related
to the presentation of <literals> which in this case are (nearly) undistinguishable
from normal text.

Actually, we should strive for a more general solution here, since there are similar
issues in the return values section, and possibly in translations.

git-svn-id: https://svn.php.net/repository/phpdoc/en/trunk@344926 c90b9560-bf6c-de11-be94-00142212c4b1